create HostBuilder to be able to set up the application's services;
